put /rendezvous stuff in its own file
